Add preliminary work on flip tracker
The tracker will count the number of times the sign of the angular velocity changes for each axis. It currently does not work – this is a preliminary commit. (It currently tracks when the sign of the angular acceleration of the intermediate axis changes. I will work towards tracking when the sign of the angular velocity of all axes changes.)